Create Online phone shop display
In this project we will first add and then retrieve the data. This will be two calls to the database. The first page will show all the phones (name) and a link 'more detail', and onced the user clicks on the link all the phone details will be shown this project will not show the UI part but will write corresponding queries to get the correct data